compose_actions: Extract compose_reply module.
