+++ web/tests/compose_actions.test.js
@@ -64,15 +64,16 @@ const people = zrequire("people");
 const compose_setup = zrequire("compose_setup");
 const compose_state = zrequire("compose_state");
 const compose_actions = zrequire("compose_actions");
+const compose_reply = zrequire("compose_reply");
 const message_lists = zrequire("message_lists");
 const stream_data = zrequire("stream_data");
 const compose_recipient = zrequire("compose_recipient");
 
 const start = compose_actions.start;
 const cancel = compose_actions.cancel;
-const respond_to_message = compose_actions.respond_to_message;
-const reply_with_mention = compose_actions.reply_with_mention;
-const quote_and_reply = compose_actions.quote_and_reply;
+const respond_to_message = compose_reply.respond_to_message;
+const reply_with_mention = compose_reply.reply_with_mention;
+const quote_and_reply = compose_reply.quote_and_reply;
 
 function assert_visible(sel) {
     assert.ok($(sel).visible());
