+++ web/tests/compose_ui.test.js
@@ -25,7 +25,7 @@ const people = zrequire("people");
 const user_status = zrequire("user_status");
 const hash_util = mock_esm("../src/hash_util");
 const channel = mock_esm("../src/channel");
-const compose_actions = zrequire("compose_actions");
+const compose_reply = zrequire("compose_reply");
 const message_lists = zrequire("message_lists");
 const text_field_edit = mock_esm("text-field-edit");
 
@@ -352,7 +352,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
     let quote_text = "Testing caret position";
     override_with_quote_text(quote_text);
     set_compose_content_with_caret("hello %there"); // "%" is used to encode/display position of focus before change
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     success_function({
         raw_content: quote_text,
@@ -367,7 +367,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
         assert.equal(syntax, "translated: [Quoting…]\n\n");
     });
     set_compose_content_with_caret("%hello there");
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     quote_text = "Testing with caret initially positioned at 0.";
     override_with_quote_text(quote_text);
@@ -375,7 +375,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
         raw_content: quote_text,
     });
 
-    override_rewire(compose_actions, "respond_to_message", () => {
+    override_rewire(compose_reply, "respond_to_message", () => {
         // Reset compose state to replicate the re-opening of compose-box.
         textarea_val = "";
         textarea_caret_pos = 0;
@@ -387,7 +387,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
     // If the compose-box is close, or open with no content while
     // quoting a message, the quoted message should be placed
     // at the beginning of compose-box.
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     quote_text = "Testing with compose-box closed initially.";
     override_with_quote_text(quote_text);
@@ -402,7 +402,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
     // newlines), the compose-box should re-open and thus the quoted
     // message should start from the beginning of compose-box.
     set_compose_content_with_caret("  \n\n \n %");
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     quote_text = "Testing with compose-box containing whitespaces and newlines only.";
     override_with_quote_text(quote_text);
@@ -419,7 +419,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
         assert.equal(syntax, "\ntranslated: [Quoting…]\n");
     });
     set_compose_content_with_caret("1st line\n%\n2nd line");
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     quote_text = "Testing with caret on a new line between 2 lines of text.";
     override_with_quote_text(quote_text);
@@ -436,7 +436,7 @@ run_test("quote_and_reply", ({override, override_rewire}) => {
         assert.equal(syntax, "translated: [Quoting…]");
     });
     set_compose_content_with_caret("lots of\n\n\n\n%\n\n\nnewlines");
-    compose_actions.quote_and_reply();
+    compose_reply.quote_and_reply();
 
     quote_text = "Testing with caret on a new line between many empty newlines.";
     override_with_quote_text(quote_text);
